+++ CHANGES.md
@@ -5,6 +5,7 @@
 - Detection of foreign default gateway VPN
 - Fixes: #214, #217, #218
 - GUI refinements
+- "Renew Session" button and session expiration notification timings revised
 
 
 ## [3.3.6](https://github.com/Amebis/eduVPN/compare/3.3.5...3.3.6) (2022-11-23)